Separate CommonServices into specific Services
